refactor(domain): rename camelCase source files to kebab-case
Bring all packages/domain/src files into compliance with the
project ls-lint convention. The exported symbol names are
unchanged, so external consumers (apps/api, apps/web) need no
import updates.

Refs #162
