+++ packages/domain/src/artifact-plan-validation.ts
