+++ packages/domain/src/artifact-plan.ts
