+++ packages/domain/src/representations/collection-json/teams.ts
@@ -21,10 +21,10 @@ import {
   type Template,
 } from '@genshin/collection-json';
 
-import type { ArtifactPlan } from '../../artifactPlan.js';
-import type { CollectionTeam, CollectionTeamMembers } from '../../collectionTeam.js';
-import { assertCollectionTeam, MAX_TEAM_MEMBERS } from '../../collectionTeam.js';
-import type { CollectionTeamMember } from '../../collectionTeamMember.js';
+import type { ArtifactPlan } from '../../artifact-plan.js';
+import type { CollectionTeam, CollectionTeamMembers } from '../../collection-team.js';
+import { assertCollectionTeam, MAX_TEAM_MEMBERS } from '../../collection-team.js';
+import type { CollectionTeamMember } from '../../collection-team-member.js';
 
 const TEAM_TEMPLATE: Template = {
   data: [
